+++ src/main/java/com/aaax/server/App.java
@@ -3,11 +3,15 @@ package com.aaax.server;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.boot.autoconfigure.kafka.KafkaAutoConfiguration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 
-@SpringBootApplication(scanBasePackages = {"com.aaax.core", "com.aaax.server"})
+@SpringBootApplication(
+        scanBasePackages = {"com.aaax.core", "com.aaax.server"},
+        exclude = KafkaAutoConfiguration.class
+)
 @EnableJpaAuditing
 @EnableTransactionManagement
 @EnableMethodSecurity // -> for use @PreAuthorize, @PreAuthorize("permitAll()"),