+++ src/main/java/com/aaax/server/listener/UserDeletedNotificationListener.java
@@ -13,6 +13,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.clients.admin.NewTopic;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.kafka.support.Acknowledgment;
@@ -21,6 +22,7 @@ import org.springframework.stereotype.Component;
 import static com.aaax.core.kafka.enu.KafkaTopic.USER_DELETED;
 
 @Component
+@ConditionalOnProperty(prefix = "aaax.kafka", name = "enabled", havingValue = "true")
 @Slf4j
 @RequiredArgsConstructor
 public class UserDeletedNotificationListener extends BaseListener {
@@ -28,7 +30,7 @@ public class UserDeletedNotificationListener extends BaseListener {
     private final KafkaUtil kafkaUtil;
     private final DiscordApiClient discordApiClient;
     private final UserManagementUseCase userManagementUseCase;
-    @Value("${config.system-invoker}")
+    @Value("${aaax.config.system-invoker}")
     protected String systemInvoker;
     @Value("${ext.api.client.discord.webhookId:}")
     private String webhookId;